Move GitopsGpgKeypair into CpEnv namespace
