+++ lib/cp_env.rb
@@ -1,8 +1,10 @@
 require "bundler/setup"
-require "kubeclient"
-require "open3"
-require "open-uri"
 require "aws-sdk-s3"
+require "kubeclient"
+require "open-uri"
+require "open3"
+require 'base64'
+require 'json'
 
 class CpEnv
 end
@@ -12,7 +14,7 @@ require File.join(File.dirname(__FILE__), "cp_env", "pipeline")
 require File.join(File.dirname(__FILE__), "cp_env", "terraform")
 require File.join(File.dirname(__FILE__), "cp_env", "namespace_deleter")
 require File.join(File.dirname(__FILE__), "cp_env", "kubeconfig")
+require File.join(File.dirname(__FILE__), "cp_env", "gitops_gpg_keypair")
 
 # TODO: Move gpg stuff into CpEnv
-require File.join(File.dirname(__FILE__), "gitops_gpg_keypair")
 require File.join(File.dirname(__FILE__), "gpg_keypair")
